Add initial version of OOPS team etiquette (#1)
New members to the team will benefit from knowing what the existing team considers to be good office etiquette. This should reduce the chance of conflict or resentment.

This change adds an initial version of a guide on etiquette. It also includes feedback from @Qbtron22 regards the morale boosting benefit of gossiping.